chore: add vendors package, reduce duplications
