+++ apps/feature-flags/app/api/v1/jobs/send-email/route.tsx
@@ -5,15 +5,10 @@ import { Receiver } from "@upstash/qstash";
 import React, { ElementType } from "react";
 import type { SendEmailPayload } from "@basestack/vendors";
 // Email
-import {
-  sendEmail,
-  NewSubmissionEmailTemplate,
-  WelcomeEmailTemplate,
-} from "@basestack/emails";
+import { sendEmail, WelcomeEmailTemplate } from "@basestack/emails";
 import { render } from "@react-email/render";
 
 const templateList: { [key: string]: ElementType } = {
-  "new-submission": NewSubmissionEmailTemplate,
   welcome: WelcomeEmailTemplate,
 };
 
@@ -22,10 +17,10 @@ export const { POST } = serve<SendEmailPayload>(
     const { to, subject, template, props } = context.requestPayload;
 
     console.info(
-      `Job: Basestack Forms - Send Email - Preparing to send email to ${to} with subject: ${subject}`,
+      `Job: Basestack Feature Flags - Send Email - Preparing to send email to ${to} with subject: ${subject}`,
     );
     console.info(
-      `Job: Basestack Forms - Send Email - Email with the template ${template} with props: ${JSON.stringify(props)}`,
+      `Job: Basestack Feature Flags - Send Email - Email with the template ${template} with props: ${JSON.stringify(props)}`,
     );
 
     await context.run("send-email-step", async () => {
@@ -47,7 +42,7 @@ export const { POST } = serve<SendEmailPayload>(
       );
 
       console.info(
-        "Job: Basestack Forms - Send Email - ✨ Email sent successfully! ✨",
+        "Job: Basestack Feature Flags - Send Email - ✨ Email sent successfully! ✨",
       );
     });
   },
@@ -63,7 +58,7 @@ export const { POST } = serve<SendEmailPayload>(
       failHeaders,
     }) => {
       console.error(
-        `Job: Basestack Forms - Send Email - status = ${JSON.stringify(failStatus)} response = ${JSON.stringify(failResponse)} headers = ${JSON.stringify(failHeaders)} context = ${JSON.stringify(context)} `,
+        `Job: Basestack Feature Flags - Send Email - status = ${JSON.stringify(failStatus)} response = ${JSON.stringify(failResponse)} headers = ${JSON.stringify(failHeaders)} context = ${JSON.stringify(context)} `,
       );
     },
   },
