+++ apps/feature-flags/package.json
@@ -34,6 +34,7 @@
     "@basestack/hooks": "*",
     "@basestack/ui": "*",
     "@basestack/utils": "*",
+    "@basestack/vendors": "*",
     "@floating-ui/react": "^0.27.5",
     "@hookform/resolvers": "^4.1.3",
     "@lemonsqueezy/lemonsqueezy.js": "^4.0.0",
