--- apps/feature-flags/server/auth/config.ts
@@ -1,118 +1,0 @@
-// // Adapters
-import { PrismaAdapter } from "@auth/prisma-adapter";
-// Types
-import { type DefaultSession, type NextAuthConfig } from "next-auth";
-import { Role } from ".prisma/client";
-import type { Provider } from "next-auth/providers";
-// Libs
-import { prisma } from "server/db";
-// Providers
-import GitHubProvider from "next-auth/providers/github";
-import GoogleProvider from "next-auth/providers/google";
-
-declare module "next-auth" {
-  interface Session extends DefaultSession {
-    user: {
-      id: string;
-      name: string;
-      email: string;
-      image?: string | null;
-      role?: Role;
-    } & DefaultSession["user"];
-  }
-
-  interface User {
-    role?: Role;
-  }
-}
-
-const providers: Provider[] = [
-  GitHubProvider({
-    clientId: process.env.GITHUB_CLIENT_ID!,
-    clientSecret: process.env.GITHUB_CLIENT_SECRET!,
-    profile(profile) {
-      return {
-        id: profile.id.toString(),
-        name: profile.name || profile.login,
-        gh_username: profile.login,
-        email: profile.email,
-        image: profile.avatar_url,
-      };
-    },
-  }),
-  GoogleProvider({
-    clientId: process.env.GOOGLE_CLIENT_ID!,
-    clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    authorization: {
-      params: {
-        prompt: "consent",
-        access_type: "offline",
-        response_type: "code",
-      },
-    },
-    profile: (profile) => {
-      return {
-        id: profile.sub,
-        name: profile.name,
-        email: profile.email,
-        image: profile.picture,
-      };
-    },
-  }),
-];
-
-export const providerMap = providers
-  .map((provider) => {
-    if (typeof provider === "function") {
-      const providerData = provider();
-      return { id: providerData.id, name: providerData.name };
-    } else {
-      return { id: provider.id, name: provider.name };
-    }
-  })
-  .filter((provider) => provider.id !== "credentials");
-
-export const authConfig: NextAuthConfig = {
-  adapter: PrismaAdapter(prisma),
-  secret: process.env.AUTH_SECRET,
-  pages: {
-    signIn: "/auth/sign-in",
-    signOut: "/auth/sign-out",
-    error: "/auth/error", // Error code passed in query string as ?error=
-    verifyRequest: "/auth/verify-request", // (used for check email message)
-  },
-  providers,
-  callbacks: {
-    session: ({ session, user }) => ({
-      ...session,
-      user: {
-        ...session.user,
-        id: user.id,
-        username: user.name,
-        role: user.role,
-      },
-    }),
-  },
-  events: {
-    async signIn(message) {
-      if (message.isNewUser) {
-        const email = message.user.email as string;
-        const user = await prisma.user.findUnique({
-          where: { email },
-          select: {
-            name: true,
-            createdAt: true,
-          },
-        });
-
-        const isNewUserRecently =
-          user?.createdAt &&
-          new Date(user.createdAt).getTime() > Date.now() - 10000;
-
-        if (isNewUserRecently) {
-          console.info("New user signed in:", message.user.email);
-        }
-      }
-    },
-  },
-};
