+++ apps/forms/app/api/v1/jobs/check-subscription/route.ts
@@ -1,89 +1,36 @@
-// UpStash
-import { serve } from "@upstash/workflow/nextjs";
-import { Receiver } from "@upstash/qstash";
 // Prisma
 import { prisma } from "server/db";
 // Utils
 import dayjs from "dayjs";
-import { PlanTypeId, config, SubscriptionEvent } from "@basestack/utils";
+import { config } from "@basestack/utils";
+// Vendors
+import { qstash } from "@basestack/vendors";
 
 const { getFormPlanLimitsDefaults } = config.plans;
 
-//  cron: "0 19 * * *", // Run every day at 7 PM
+export const { POST } = qstash.jobs.CheckSubscriptionJob({
+  product: "Forms",
+  getSubs: async () => {
+    const res = await prisma.subscription.findMany();
 
-export const { POST } = serve(
-  async (context) => {
-    console.info(
-      "Job: Check Forms Subscriptions - Received the scheduled event",
-    );
-
-    await context.run("check-users-subscriptions-step", async () => {
-      const subs = await prisma.subscription.findMany();
-
-      for (const sub of subs) {
-        const billingCycle = dayjs(sub.billingCycleStart);
-        const today = dayjs();
-
-        // Check if today is the billing cycle start date or later
-        const isOverdue =
-          today.isAfter(billingCycle.startOf("day")) ||
-          today.isSame(billingCycle.startOf("day"));
-
-        console.info(
-          `Job: Check Forms Subscriptions - User with ID: ${sub.userId} subscription billingCycle is ${billingCycle.format("YYYY-MM-DD")} and is overdue: ${isOverdue}`,
-        );
-
-        // Check if it's time to update the subscription
-        if (isOverdue) {
-          let payload = {};
-
-          if (sub.cancelled || sub.paused) {
-            console.info(
-              `Job: Check Forms Subscriptions - User with ID ${sub.userId} has an cancelled or paused subscription`,
-            );
-
-            payload = {
-              planId: PlanTypeId.FREE,
-              event: SubscriptionEvent.SUBSCRIPTION_CANCELLED,
-              subscriptionId: "",
-            };
-          }
-
-          const response = await prisma.subscription.update({
-            where: {
-              id: sub.id,
-            },
-            data: {
-              billingCycleStart: dayjs(billingCycle)
-                .add(1, "month")
-                .toISOString(),
-              ...getFormPlanLimitsDefaults(),
-              ...payload,
-            },
-          });
-
-          console.info(
-            `Job: Check Forms Subscriptions - User ${sub.userId} subscription updated successfully`,
-            response,
-          );
-        }
-      }
+    return res.map(({ id, billingCycleStart, userId, cancelled, paused }) => ({
+      id,
+      billingCycleStart,
+      userId,
+      cancelled: cancelled ?? false,
+      paused: paused ?? false,
+    }));
+  },
+  onSuccess: (id, billingCycle, payload) => {
+    return prisma.subscription.update({
+      where: {
+        id,
+      },
+      data: {
+        billingCycleStart: dayjs(billingCycle).add(1, "month").toISOString(),
+        ...getFormPlanLimitsDefaults(),
+        ...payload,
+      },
     });
   },
-  {
-    receiver: new Receiver({
-      currentSigningKey: process.env.QSTASH_CURRENT_SIGNING_KEY!,
-      nextSigningKey: process.env.QSTASH_NEXT_SIGNING_KEY!,
-    }),
-    failureFunction: async ({
-      context,
-      failStatus,
-      failResponse,
-      failHeaders,
-    }) => {
-      console.error(
-        `Job: Check Forms Subscriptions - status = ${JSON.stringify(failStatus)} response = ${JSON.stringify(failResponse)} headers = ${JSON.stringify(failHeaders)} context = ${JSON.stringify(context)} `,
-      );
-    },
-  },
-);
+});
