+++ apps/forms/package.json
@@ -34,6 +34,7 @@
     "@basestack/hooks": "*",
     "@basestack/ui": "*",
     "@basestack/utils": "*",
+    "@basestack/vendors": "*",
     "@hookform/resolvers": "^4.1.3",
     "@lemonsqueezy/lemonsqueezy.js": "^4.0.0",
     "@prisma/client": "^6.4.1",
