+++ apps/forms/server/auth/index.ts
@@ -2,8 +2,40 @@
 import NextAuth from "next-auth";
 // Utils
 import { cache } from "react";
+// DB
+import { prisma } from "server/db";
+// Vendors
+import { PrismaAdapter } from "@auth/prisma-adapter";
+import { auth as authVendor } from "@basestack/vendors";
 
-import { authConfig } from "./config";
+const authConfig = authVendor.createAuthConfig({
+  product: "Forms",
+  adapter: PrismaAdapter(prisma),
+  getUser: async (email) => {
+    const user = await prisma.user.findUnique({
+      where: { email },
+      select: {
+        name: true,
+        createdAt: true,
+      },
+    });
+
+    return {
+      name: user?.name ?? "User",
+      createdAt: user?.createdAt ?? new Date(),
+    };
+  },
+  content: {
+    emails: {
+      welcome: {
+        title: "Welcome to Basestack Forms",
+        description:
+          "Welcome to Basestack Forms, the platform that elevates your website with powerful, customizable forms.",
+        link: "https://forms.basestack.co",
+      },
+    },
+  },
+});
 
 const { auth: uncachedAuth, handlers, signIn, signOut } = NextAuth(authConfig);
