--- apps/forms/server/auth/types.ts
@@ -1,12 +1,0 @@
-export type ProviderType = "oauth" | "email" | "credentials";
-
-export interface ProviderOptions {
-  id: string;
-  name: string;
-  type: ProviderType;
-  options?: Record<string, unknown>;
-}
-
-export interface Provider {
-  [key: string]: ProviderOptions;
-}
