+++ packages/emails/src/templates/invite.tsx
@@ -116,7 +116,7 @@ export const InviteEmailTemplate = ({
             Help & Support
           </Link>
           <Text style={{ ...Styles.ParagraphMuted, fontSize: "14px" }}>
-            © Basestack 2024
+            © Basestack {new Date().getFullYear()}. All rights reserved.
           </Text>
         </Container>
       </Body>
