+++ packages/vendors/jest.config.js
@@ -1,0 +1,6 @@
+const common = require("@basestack/config/jest.base.config");
+
+module.exports = {
+  ...common,
+  setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
+};
