+++ packages/vendors/jest.setup.js
@@ -1,0 +1,3 @@
+const common = require("@basestack/config/jest.setup.js");
+
+module.exports = common;
