+++ packages/vendors/src/auth/index.ts
@@ -1,0 +1,3 @@
+import { providerMap, createAuthConfig } from "./config";
+
+export { providerMap, createAuthConfig };
