+++ packages/vendors/src/auth/types.ts
