+++ packages/vendors/src/cf/ai.ts
