+++ packages/vendors/src/lemonsqueezy/index.ts
@@ -1,0 +1,7 @@
+import { createSubscriptionEvent } from "./webhooks/subscriptionEvent";
+
+const webhook = {
+  createSubscriptionEvent,
+};
+
+export { webhook };
