+++ packages/vendors/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "@basestack/tsconfig/react-library.json",
+  "compilerOptions": {
+    "types": ["node", "jest", "@testing-library/jest-dom", "react"],
+    "lib": ["ES2015", "dom"],
+    "baseUrl": ".",
+    "noEmit": true
+  },
+  "include": ["."],
+  "exclude": ["dist", "build", "node_modules"]
+}
