+++ turbo.json
@@ -13,7 +13,17 @@
     "GOOGLE_CLIENT_SECRET",
     "AUTH_SECRET",
     "AUTH_URL",
-    "NEXT_PUBLIC_APP_MODE"
+    "NEXT_PUBLIC_APP_MODE",
+    "UPSTASH_WORKFLOW_URL",
+    "QSTASH_TOKEN",
+    "QSTASH_URL",
+    "QSTASH_CURRENT_SIGNING_KEY",
+    "QSTASH_NEXT_SIGNING_KEY",
+    "LEMONSQUEEZY_STORE_ID",
+    "LEMONSQUEEZY_SIGNATURE_SECRET",
+    "LEMONSQUEEZY_API_KEY",
+    "CLOUDFLARE_ACCOUNT_ID",
+    "CLOUDFLARE_API_TOKEN"
   ],
   "tasks": {
     "clean": {
@@ -41,36 +51,12 @@
     "@basestack/feature-flags#deploy": {
       "dependsOn": ["^build", "lint"],
       "outputs": ["dist/**", ".next/**"],
-      "env": [
-        "DATABASE_URL",
-        "DIRECT_DATABASE_URL",
-        "UPSTASH_WORKFLOW_URL",
-        "QSTASH_TOKEN",
-        "QSTASH_URL",
-        "QSTASH_CURRENT_SIGNING_KEY",
-        "QSTASH_NEXT_SIGNING_KEY",
-        "LEMONSQUEEZY_STORE_ID",
-        "LEMONSQUEEZY_SIGNATURE_SECRET",
-        "LEMONSQUEEZY_API_KEY"
-      ]
+      "env": ["DATABASE_URL", "DIRECT_DATABASE_URL"]
     },
     "@basestack/forms#deploy": {
       "dependsOn": ["^build", "lint"],
       "outputs": ["dist/**", ".next/**"],
-      "env": [
-        "DATABASE_URL",
-        "DIRECT_DATABASE_URL",
-        "CLOUDFLARE_ACCOUNT_ID",
-        "CLOUDFLARE_API_TOKEN",
-        "UPSTASH_WORKFLOW_URL",
-        "QSTASH_TOKEN",
-        "QSTASH_URL",
-        "QSTASH_CURRENT_SIGNING_KEY",
-        "QSTASH_NEXT_SIGNING_KEY",
-        "LEMONSQUEEZY_STORE_ID",
-        "LEMONSQUEEZY_SIGNATURE_SECRET",
-        "LEMONSQUEEZY_API_KEY"
-      ]
+      "env": ["DATABASE_URL", "DIRECT_DATABASE_URL"]
     },
     "@basestack/landing-page#deploy": {
       "dependsOn": [
